portability/warning fixes
git-svn-id: https://dev.csoft.net/agar/trunk@6477 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ core/load_netbuf.c
@@ -30,7 +30,6 @@
 #include <core/core.h>
 
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include <errno.h>
 
@@ -76,7 +75,7 @@ AG_NetbufRead(void *ptr, size_t size, size_t nmemb, AG_Netbuf *buf)
 __inline__ long
 AG_NetbufReadE(void *ptr, size_t size, size_t nmemb, AG_Netbuf *buf)
 {
-	return (fread(ptr, size, nmemb, buf->file));
+	return (long)fread(ptr, size, nmemb, buf->file);
 }
 
 void
